scripts: implement function that searches value in values set
We have implemented 'is_value_in_set' function that checks whether
value (first argument) presents in the set of values
(second argument). The function is more generic than the already
existing ool_contains and may be used for searching for a value
in arbitrary set. The ool_contains is modified to use
is_value_in_set.

OL-Redmine-Id: 11959
Signed-off-by: Pavel Liulchak <pavel.liulchak@oktetlabs.ru>
Reviewed-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>