scripts: disable vlan if af_xdp over vlan on Intel/Mellanox
In case of aggregation + netns_iut we have to do tests
either with macvlan or ipvlan thus we replace
vlan on macvlan in case of af_xdp. In all other
af_xdp cases we have to just delete the vlan option,
we do this within the patch and simplify the if-else
conditions.

Fixes: 0a3c24d7215a ("scripts: disable vlan if test af_xdp over vlan on Intel/Mellanox")
OL-Redmine-Id: 11959
Signed-off-by: Pavel Liulchak <pavel.liulchak@oktetlabs.ru>
Reviewed-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>