sockopts/tcp_info_retransmits: fix sleeping value
When catching a TCP retransmit, the test should sleep for RTO*3 time.
It ensures that one and only one retransmit will happen suring sleep.
TAPI_WAIT_NETWORK is completely different and should not be used for
such a goal.

As a side effect the patch increases the sleeping timeout, making the
test to behave on Linux.

OL-Redmine-Id: 13321
Signed-off-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>
Reviewed-by: Damir Mansurov <damir.mansurov@oktetlabs.ru>