trc: fix TRC records to match schema
As the schema for TRC DB is now up to date
in TE, some minor issues are found and are
fixed by this patch. No behaviour changes
are expected: the changes are either fixes
of XML syntax errors or removal of attributes
that are not processed by TRC in any way.

OL-Redmine-Id: 14147
Signed-off-by: Artem Andreev <artem.andreev@oktetlabs.ru>
Reviewed-by: Alexandra Kossovsky <alexandra.kossovsky@oktetlabs.ru>